requirements.txt
# Data collection
apify-client==2.5.0
requests==2.32.3

# Database
supabase==2.28.3

# Environment
python-dotenv==1.2.1

# Image processing & pHash
Pillow==10.3.0
imagehash==4.3.1

# Data & ML
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4

# Sentiment analysis
transformers==4.40.0
torch==2.3.0

# Price forecasting
prophet==1.1.5

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# OpenClaw workspace — manifest parsing only.
# Agents / Telegram / LLM are provided by the OpenClaw Gateway itself (not Python).
PyYAML==6.0.2
